CHAT-45: Add api backend cors
