+++ chatmeleon-server/.env.example
@@ -3,4 +3,6 @@ DATABASE_URL= " "
 CHAT_TOKEN_SECRET =
 
 # Comma-seperated origins
-SOCKET_IO_CORS_ORIGINS = "http://localhost:3000,http://localhost:5000"
+# E.g, "http://localhost:3000,http://localhost:5000,https://front-end:443"
+API_CORS_ORIGINS = "http://localhost:3000"
+SOCKET_IO_CORS_ORIGINS = "http://localhost:3000"
