+++ chatmeleon-server/src/index.ts
@@ -19,6 +19,8 @@ import { conversationRoute } from "./routes/conversation-route";
 import { userRoute } from "./routes/user-route";
 import { createServer } from "http";
 import { io } from "./libs/socket.io";
+import cors from "cors";
+import { parseCorsOrigin } from "./utils";
 
 const app = express();
 const httpServer = createServer(app);
@@ -26,14 +28,19 @@ io.attach(httpServer);
 
 const port = process.env.PORT;
 
+app.use(
+  cors({
+    origin: parseCorsOrigin(process.env.API_CORS_ORIGINS),
+  })
+);
+app.use(express.json());
+
 // Serve Swagger documentation
 app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 
 app.get("/", (req: Request, res: Response) => {
   res.send("Backend server is running.");
 });
-
-app.use(express.json());
 
 app.use(messageRoute);
 app.use(conversationRoute);
