+++ chatmeleon-server/src/libs/socket.io/index.ts
@@ -1,16 +1,5 @@
 import { Server } from "socket.io";
-
-function parseCorsOrigin(value: string | undefined) {
-  if (value === undefined) return;
-
-  const origins = value.split(",");
-
-  if (origins.length <= 1) {
-    return origins[0];
-  }
-
-  return origins;
-}
+import { parseCorsOrigin } from "../../utils";
 
 export const io = new Server({
   cors: {
